ai project/cp3/Proj 3/requirements.txt
Django==5.0.2
google-generativeai==0.3.2
python-dotenv==1.0.1
requests==2.31.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10 